added the UI class to increase OOP
